[PT-5597] (1) Fix test for `CatalogBase` (#673)
* fix test for CatalogBase

* improving detail of test

* first round of feedback

* first round of feedback

* updating changelog date

* updating changelog date

* first round of feedback

+++ up42/catalog.py
@@ -45,7 +45,7 @@ class CatalogBase:
         self,
         order_parameters: order.OrderParams,
         track_status: bool = False,
-        report_time: int = 120,
+        report_time: float = 120,
     ) -> order.Order:
         """
         Place an order.
